cpu: BPU support for surprise branches
To support the decoupled front-end the BPU must have the
ability to add a placeholder for branches that are not in
the BTB (surprise branches). In this case the front-end
cannot detect the branch and make a prediction. However,
to roleback in case of a misprediction the BPU needs a
history element.
This function allows the to add such element after pre-
decode. It does NOT make a prediction or alter the BPU
state.
TAGE's global history management is update to support
roleback of surprise branches.

Change-Id: I4446f5f29908d75315ecde1eca26986b21324896
Signed-off-by: David Schall <david.schall@ed.ac.uk>