cpu: FDP fix for non-micro coded branches
As x86 implements come complex instructions like string
copy the decoupled front-end checksif an instruction is
the last micro-op. However,arm does not micro code
branches and correspondingly does not set the
`IsLastMicroop` flag and misleads the decoupling for Arm.
This fix adds a check whether an instruction is micro
coded in the first place.

Change-Id: I93522c6642b240479b8795419c5530bd41404df5